# lib/api-zod/src/generated/types/detailedHealth.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Teacher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DetailedHealth {
  status: string;
  version: string;
  llm_provider: string;
  llm_status: string;
  chromadb_status: string;
  uptime_seconds: number;
  total_jobs: number;
  active_jobs: number;
}